chunk_store: widen cs_nblocks to uint64_t
cs_nblocks was uint32_t but chunk_store_write and chunk_store_lookup
accept uint64_t offsets.  The grow function capped at 1M entries so
this was safe in practice, but the type mismatch meant offsets above
UINT32_MAX could silently bypass bounds checks after a truncating
cast.

+++ lib/nfs4/include/nfs4/chunk_store.h
@@ -43,7 +43,7 @@ struct chunk_block {
  */
 struct chunk_store {
 	struct chunk_block *cs_blocks;
-	uint32_t cs_nblocks; /* allocated entries */
+	uint64_t cs_nblocks; /* allocated entries */
 };
 
 /*
